Extract context setup
