+++ client/src/app.js
@@ -1,55 +1,16 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Router } from '@reach/router';
 import { MainView, DetailView } from './views';
 import { ThemeProvider, AppProvider } from './contexts';
 import apolloClient from './helpers/apollo-client';
 import { ApolloProvider } from 'react-apollo';
 
-const theme = {
-  primary: '#002B56',
-  buttonText: '#fff',
-  open: '#00E8A4',
-  closed: '#FF3548',
-  fontLight: '#757575',
-  borderLight: '#C8C8C8'
-};
-
 const App = () => {
-  const [openNow, setOpenNow] = useState(false);
-  const [pricesSelected, setPricesSelected] = useState([]);
-  const [categoriesSelected, setCategoriesSelected] = useState([]);
-
-  // handle toggling multiple values
-  // by adding/removing them from an array in local state
-  const toggleSelected = (value, values, stateSetter) =>
-    values.includes(value) ?
-    // remove value
-    stateSetter([...values.filter(val => val !== value)]) :
-    // add value
-    stateSetter([...values, value]);
-  
-
-  // setup app context
-  // this is mostly used by the filter component
-  const context = {
-    openNow,
-    pricesSelected,
-    categoriesSelected,
-    setCategoriesSelected,
-    setPricesSelected,
-    toggleOpenNow: () =>
-      setOpenNow(!openNow),
-    togglePriceSelect: (price) =>
-      toggleSelected(price, pricesSelected, setPricesSelected),
-    toggleCategorySelect: (category) =>
-      toggleSelected(category, categoriesSelected, setCategoriesSelected)
-  };
-
   return (
     <div className='app'>
       <ApolloProvider client={apolloClient}>
-        <AppProvider value={context}>
-          <ThemeProvider value={theme}>
+        <AppProvider>
+          <ThemeProvider>
             <Router>
               <DetailView path="/:slug/:id" />
               <MainView path="/" />
